Move directory publish integration tests into `ivy` and `maven` subprojects
