+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/DirectoryOutputArtifactIntegrationTest.groovy
@@ -18,8 +18,6 @@ package org.gradle.integtests.resolve
 import groovy.transform.NotYetImplemented
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 
-import static org.hamcrest.core.StringContains.containsString
-
 class DirectoryOutputArtifactIntegrationTest extends AbstractIntegrationSpec {
 
     def "can attach a directory as output of a configuration"() {
@@ -218,68 +216,6 @@ class DirectoryOutputArtifactIntegrationTest extends AbstractIntegrationSpec {
         executedAndNotSkipped ':b:compileJava'
         notExecuted ':b:jar'
 
-    }
-
-    def "fails gracefully if trying to publish a directory with ivy"() {
-
-        given:
-        file('someDir/a.txt') << 'some text'
-        buildFile << """
-
-        apply plugin: 'base'
-
-        configurations {
-            archives
-        }
-
-        artifacts {
-            archives file("someDir")
-        }
-
-        """
-
-        when:
-        fails 'uploadArchives'
-
-        then:
-        failure.assertHasCause "Could not publish configuration 'archives'"
-        failure.assertThatCause(containsString('Cannot publish a directory'))
-
-    }
-
-    def "fails gracefully if trying to publish a directory with Maven"() {
-
-        given:
-        file('someDir/a.txt') << 'some text'
-        buildFile << """
-
-        apply plugin: 'base'
-        apply plugin: 'maven'
-
-        uploadArchives {
-            repositories {
-                mavenDeployer {
-                    repository(url: "${mavenRepo.uri.toURL()}")
-                }
-            }
-        }
-
-        configurations {
-            archives
-        }
-
-        artifacts {
-            archives file("someDir")
-        }
-
-        """
-
-        when:
-        fails 'uploadArchives'
-
-        then:
-        failure.assertHasCause "Could not publish configuration 'archives'"
-        failure.assertThatCause(containsString('Cannot publish a directory'))
     }
 
     def "can avoid building a jar when compiling against another project with transitive dependencies"() {
